Status-curing items remove stat modifiers
Using an item that cures a Pokémon's status conditions will reset their stats back to their neutral, unmodified values. This is done to remove the stat changes caused by burn and paralysis. However, the modifications from stat changing moves, as well as badge boosts, do not get reapplied. In addition, the number of stat stages does not get updated, which could cause stat changing moves to not behave properly if a stat was at either the +6 or -6 cap.